+++ queng_infra/postgres/pg_imdb/src/model/integration_message_config_type/integration_message_config_type_conversion.rs
@@ -18,7 +18,7 @@ impl MessageConfig {
         IntegrationMessageConfig::new(
             self.id as u16,
             self.version as u16,
-            ExchangeID::from(self.exchange_id as u16),
+            ExchangeID::from(self.exchange_id),
         )
     }
 }
